RDCC-2961: Moved EmailConfiguration to configuration directory
