+++ src/main/java/uk/gov/hmcts/reform/juddata/camel/util/JrdDataIngestionLibraryRunner.java
@@ -16,7 +16,7 @@ import uk.gov.hmcts.reform.data.ingestion.camel.service.IEmailService;
 import uk.gov.hmcts.reform.data.ingestion.camel.service.dto.Email;
 import uk.gov.hmcts.reform.juddata.camel.servicebus.TopicPublisher;
 import uk.gov.hmcts.reform.juddata.client.IdamClient;
-import uk.gov.hmcts.reform.juddata.email.config.EmailConfiguration;
+import uk.gov.hmcts.reform.juddata.configuration.EmailConfiguration;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
